SVM working for our data
In order to make it work, I had to re-do the data_targets to be 1-D from 5-D. TO do that I used [1,2,3,4,5] as our classifiers for the different combinations of [1,-1,-1,-1,-1] vectors. Not sure if this is the right think to do.
RVM is so slow that I waited about 10 misn for an output and gave up.
Gaussian Process works (but slowly), but I get terrible results. Not sure why.